backend/solark_browser/requirements.txt
psycopg2-binary>=2.9      # DB adapter – required
# If you later parse .env inside Python:
python-dotenv>=1.0
